move functions back to IndexPoolUtils
