+++ pkg/pool-weighted/test/utils/WeightCalculationUtil.test.ts
@@ -57,41 +57,3 @@ export const getRandomBaseWeights = (numWeights: number): number[] => {
 
   return baseWeights;
 };
-
-export const getIntegerBetween = (min: number, max: number): number => Math.floor(Math.random() * max + min);
-
-// generates random baseWeights and fixedWeights for the case where the weights of existing tokens in a pool are changed
-export const setupAdjustTokens = (numWeights: number) => {
-  const baseWeights = getRandomBaseWeights(numWeights);
-  const numberAdjustTokens = getIntegerBetween(1, numWeights - 1);
-  const fixedWeights = new Array(numWeights).fill(0);
-
-  let residual = 1;
-  for (let i = 0; i < numberAdjustTokens; i++) {
-    const newWeight = getDecimalBetween(Number.MIN_VALUE, residual / 2);
-    fixedWeights[i] = newWeight;
-    baseWeights[i] = 0;
-    residual -= newWeight;
-  }
-
-  return { baseWeights, fixedWeights };
-};
-
-// generates random baseWeights and fixedWeights for the case where one or two new tokens are added
-export const setupNewTokens = (numWeights: number) => {
-  const baseWeights = getRandomBaseWeights(numWeights);
-
-  const numberNewTokens = getIntegerBetween(1, 2);
-  const fixedWeights = new Array(baseWeights.length).fill(0);
-  const initialWeight = 0.01;
-  for (let i = 0; i < numberNewTokens; i++) {
-    fixedWeights.push(initialWeight);
-  }
-  const numberAddedTokens = fixedWeights.length - baseWeights.length;
-
-  for (let i = 0; i < numberAddedTokens; i++) {
-    baseWeights.push(0);
-  }
-
-  return { baseWeights, fixedWeights };
-};
